Make each page a seperate component
