+++ frontend/src/Experiment.jsx
@@ -1,0 +1,18 @@
+import React, { Component } from 'react';
+import logo from './logo.png';
+import './App.css';
+
+class Experiment extends Component {
+  render() {
+    return (
+      <div>
+        <h2>Experiment</h2>
+        <p>
+          This page will be used to test out experimental UIs.
+        </p>
+      </div>
+    );
+  }
+}
+
+export default Experiment;
