+++ frontend/src/Home.test.js
@@ -1,20 +1,20 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import { shallow } from 'enzyme';
-import App from './App';
+import Home from './Home';
 
 it('renders without crashing', () => {
   const div = document.createElement('div');
-  ReactDOM.render(<App />, div);
+  ReactDOM.render(<Home />, div);
   ReactDOM.unmountComponentAtNode(div);
 });
 
 it('renders without crashing', () => {
-  shallow(<App />);
+  shallow(<Home />);
 });
 
 it('renders initial message from backend', () => {
-  const wrapper = shallow(<App />);
+  const wrapper = shallow(<Home />);
   const initialMessage = <p>The backend says: nothing.</p>;
   expect(wrapper.contains(initialMessage)).toEqual(true);
 });
