+++ frontend/src/Status.jsx
@@ -1,0 +1,18 @@
+import React, { Component } from 'react';
+import logo from './logo.png';
+import './App.css';
+
+class Status extends Component {
+  render() {
+    return (
+      <div>
+        <h2>Status</h2>
+        <p>
+          This page will contain status information about the state of the application.
+        </p>
+      </div>
+    );
+  }
+}
+
+export default Status;
